auth finished

+++ graphql/resolvers/bookings.js
@@ -1,18 +1,21 @@
 const EventModel = require("../../models/event");
 const BookingModel = require("../../models/booking");
-const {dateToString} = require('../../helpers/date');
+const { dateToString } = require("../../helpers/date");
 
 module.exports = {
   ///////////////////////////////////////////////////
   // Return all bookings if exist
-  bookings: async () => {
+  bookings: async (args, req) => {
+    if (!req.isAuth) {
+      throw new Error("Unautheticated!");
+    }
     try {
       const res = await BookingModel.find();
       return res.map((item) => {
         return {
           ...item._doc,
           createdAt: dateToString(item._doc.createdAt),
-          updatedAt: dateToString(item._doc.updatedAt)
+          updatedAt: dateToString(item._doc.updatedAt),
         };
       });
     } catch (e) {
@@ -22,30 +25,36 @@ module.exports = {
 
   ///////////////////////////////////////////////////
   // Create new Booking
-  bookEvent: async ({eId}) => {
+  bookEvent: async ({ eId }, req) => {
+    if (!req.isAuth) {
+      throw new Error("Unautheticated!");
+    }
     try {
       const event = await EventModel.findById(eId);
       const booking = new BookingModel({
-          event,
-          user: '5f5faefe4a3211399cd0ecce'
+        event,
+        user: req.userId
       });
       const res = await booking.save();
       return {
-          ...res,
-          _id: res._id,
-          createdAt: dateToString(res._doc.createdAt),
-          updatedAt: dateToString(res._doc.updatedAt)
-      }
+        ...res,
+        _id: res._id,
+        createdAt: dateToString(res._doc.createdAt),
+        updatedAt: dateToString(res._doc.updatedAt),
+      };
     } catch (e) {
-        throw e;
+      throw e;
     }
   },
 
   ///////////////////////////////////////////////////
   // Delete existing Booking
-  cancelBooking: async ({bId}) => {
-    const {event} = await BookingModel.findById(bId);
+  cancelBooking: async ({ bId }, req) => {
+    if (!req.isAuth) {
+      throw new Error("Unautheticated!");
+    }
+    const { event } = await BookingModel.findById(bId);
     await BookingModel.findByIdAndDelete(bId);
     return event;
-  }
+  },
 };
